Abstract Logic from MetadataWriter Operator for Reusability
