+++ delta/src/main/scala/io/qbeast/spark/delta/DeltaMetadataWriter.scala
@@ -54,8 +54,8 @@ import scala.collection.mutable.ListBuffer
  *
  * @param tableID
  *   the table identifier
- * @param mode
- *   SaveMode of the writeMetadata
+ * @param writeMode
+ *   WriteMode of the writeMetadata
  * @param deltaLog
  *   deltaLog associated to the table
  * @param qbeastOptions
@@ -283,7 +283,9 @@ private[delta] case class DeltaMetadataWriter(
       removeFiles: Seq[RemoveFile],
       extraConfiguration: Configuration): Seq[Action] = {
 
-    if (txn.readVersion > -1) {
+    val isNewTable = txn.readVersion == -1
+
+    if (!isNewTable) {
       // This table already exists, check if the insert is valid.
       if (saveMode == SaveMode.ErrorIfExists) {
         throw AnalysisExceptionFactory.create(s"Path '${deltaLog.dataPath}' already exists.'")
@@ -297,29 +299,29 @@ private[delta] case class DeltaMetadataWriter(
 
     val isOptimizeOperation: Boolean = tableChanges.isOptimizeOperation
 
+    val (newConfiguration, hasRevisionUpdate) = updateConfiguration(
+      txn.metadata.configuration,
+      isNewTable,
+      isOverwriteOperation,
+      tableChanges,
+      qbeastOptions)
+
     // The Metadata can be updated only once in a single transaction
     // If a new space revision or a new replicated set is detected,
     // we update everything in the same operation
-    updateQbeastMetadata(
+    updateTableMetadata(
       txn,
       schema,
       isOverwriteOperation,
       rearrangeOnly,
-      tableChanges,
-      qbeastOptions)
+      newConfiguration,
+      hasRevisionUpdate)
 
-    if (txn.readVersion < 0) {
-      // Initialize the log path
-      val fs = deltaLog.logPath.getFileSystem(sparkSession.sessionState.newHadoopConf)
+    if (isNewTable) deltaLog.createLogDirectory()
 
-      fs.mkdirs(deltaLog.logPath)
-    }
-
-    val deletedFiles = saveMode match {
-      case SaveMode.Overwrite =>
-        txn.filterFiles().map(_.remove)
-      case _ => removeFiles
-    }
+    val deletedFiles = if (isOverwriteOperation && !isNewTable) {
+      txn.filterFiles().map(_.remove)
+    } else removeFiles
 
     val allFileActions = if (rearrangeOnly) {
       addFiles.map(_.copy(dataChange = !rearrangeOnly)) ++
